refactor: 용도에 맞게 네이밍 변경
