+++ src/components/Pointillism.tsx
@@ -1,21 +1,22 @@
 import styled from "styled-components";
-import { THUMBNAIL_HEIGHT, THUMBNAIL_WIDTH } from "../../constants";
-import { useDrawThumbnail } from "../../hooks/useDrawThumbnail";
-import { getCoordinate } from "../../utils/utils";
-import { Triangle } from "./Triangle";
 
-export const Thumbnail = ({ src }: ThumbnailProps) => {
-  const { triangleArray, canvasRef, imageRef, rgba } = useDrawThumbnail();
+import { THUMBNAIL_HEIGHT, THUMBNAIL_WIDTH } from "../constants";
+import { useGetImageRgba } from "../hooks/useGetImageRgba";
+import { ThumbnailProps } from "../types";
+import { getCoordinate } from "../utils/utils";
+import { Triangle } from "./@shared/Triangle";
+
+export const Pointillism = ({ src }: ThumbnailProps) => {
+  const { componentArray, canvasRef, imageRef, rgba } = useGetImageRgba();
 
   return (
     <S.Container>
       <S.Canvas ref={canvasRef} width={THUMBNAIL_WIDTH} height={THUMBNAIL_HEIGHT}></S.Canvas>
       <S.Image src={src} ref={imageRef} />
       {rgba &&
-        triangleArray.map((el, index) => {
+        componentArray.map((el, index) => {
           const coordinate = getCoordinate(index);
           const currentPixel = Math.floor(coordinate.x + coordinate.y * THUMBNAIL_WIDTH);
-
           return (
             <Triangle
               key={index}
@@ -33,14 +34,10 @@ export const Thumbnail = ({ src }: ThumbnailProps) => {
   );
 };
 
-interface ThumbnailProps {
-  src: string;
-}
-
 const S = {
   Container: styled.div`
-    position: relative;
     overflow: hidden;
+    position: relative;
     width: 1600px;
     height: 900px;
   `,
