+++ src/hooks/useMakeRandomTriangle.ts
@@ -1,8 +1,8 @@
 import { useEffect, useState } from "react";
+import { TRIANGLE_MAX_HEIGHT, TRIANGLE_MIN_HEIGHT, TRIANGLE_WIDTH } from "../constants/index";
 import { getCoordinate } from "../utils/utils";
-import { TRIANGLE_MAX_HEIGHT, TRIANGLE_MIN_HEIGHT, TRIANGLE_WIDTH } from "./../constants/index";
 
-export const useRandomTriangle = (index: number) => {
+export const useMakeRandomTriangle = (index: number) => {
   const [height, setHeight] = useState(0);
   const [rotatePercent, setRotatePercent] = useState(0);
   const [coordinate, setCoordinate] = useState({ x: 0, y: 0 });
