+++ src/pages/ThumbnailResult.tsx
@@ -1,12 +1,17 @@
 import { useLocation } from "react-router-dom";
 import styled from "styled-components";
-import { Thumbnail } from "../components/@shared/Thumbnail";
+import { Pointillism } from "../components/Pointillism";
 
 export const ThumbnailResult = () => {
   const location = useLocation();
   const imageUrl = location.state;
 
-  return <S.Container>{imageUrl && <Thumbnail src={imageUrl} />}</S.Container>;
+  return (
+    <S.Container>
+      {imageUrl && <Pointillism src={imageUrl} />}
+      {/* {imageUrl && <Glitch src={imageUrl} />} */}
+    </S.Container>
+  );
 };
 
 const S = {
